Move cache tests to a separate file
